requirements/testing.txt
-r base.txt

factory-boy==3.2.0
pytest==6.2.2
pytest-django==4.1.0
